requirements.txt
# sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
# Remove langchain_openai
# langchain_openai
langchain_experimental
langchain_core
-e .